requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
supabase==2.4.4
python-dotenv==1.0.1
squareup==30.0.0
